Refactor update-column-type file
